# docs/swagger.yaml
basePath: /
definitions:
  domains.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  domains.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/domains.LoginResponseData'
      status:
        type: string
    type: object
  domains.LoginResponseData:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  domains.RefreshResponse:
    properties:
      data:
        $ref: '#/definitions/domains.RefreshResponseData'
      status:
        type: string
    type: object
  domains.RefreshResponseData:
    properties:
      accessToken:
        type: string
    type: object
  domains.SignupResponse:
    properties:
      data:
        $ref: '#/definitions/domains.SignupResponseData'
      message:
        type: string
      status:
        type: string
    type: object
  domains.SignupResponseData:
    properties:
      fullname:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  domains.SuccessResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:9000
info:
  contact: {}
  description: This is a Forum API
  title: Forum API
  version: "1.0"
paths:
  /authentications:
    delete:
      consumes:
      - application/json
      description: Signing User Out. Requires refresh token
      parameters:
      - description: refresh token possessed by the user
        in: body
        name: refreshToken
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domains.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
      summary: Sign Out
      tags:
      - authentication
    post:
      consumes:
      - application/json
      description: authenticate user
      parameters:
      - description: username address of the user
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: password of the user
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domains.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
      summary: Login with Username & Password
      tags:
      - authentication
    put:
      consumes:
      - application/json
      description: Generating new access token using a refresh token. Only valid refresh
        token will generate new
      parameters:
      - description: refresh token possessed by the user
        in: body
        name: refreshToken
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domains.RefreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
      summary: Refresh Authentication
      tags:
      - authentication
  /users:
    post:
      consumes:
      - application/json
      description: New user must have a unique email address
      parameters:
      - description: email address of the new user, must be unique
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: password of the new user
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: name of the new user
        in: body
        name: name
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domains.SignupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
      summary: Register A User
      tags:
      - users
securityDefinitions:
  "":
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"